+++ CHANGELOG.md
@@ -3,6 +3,9 @@
 ### Fixed
 * Re-enabled flaw history (`OSIDB-4778`)
 
+### Changed
+* Allow single CWE value only (`OSIDB-4745`)
+
 ## [2026.1.2]
 ### Fixed
 * Send partially modified Aegis-suggestions (`OSIDB-4764`)